requirements.txt
psutil==5.7.0
request==3.0.4
Django==2.2.9
django-crispy-forms==1.8.1
django-extensions==2.2.5
django-filter==2.2.0
django-import-export==2.0
django-widget-tweaks==1.4.5
djangoql==0.13.1
django-bootstrap-pagination==1.7.1
plotly==4.5.2
plotly-geo==1.0.0
pandas==1.0.1
django-queryset-csv==1.0.3

